requirements.txt
# Core Deep Learning Frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Training Optimization
unsloth>=2023.12
axolotl>=0.4.0
flash-attn>=2.4.0
triton>=2.1.0

# Data Processing
datasets>=2.15.0
tokenizers>=0.15.0
pandas>=2.1.0
numpy>=1.24.0
jsonlines>=4.0.0
pyarrow>=14.0.0

# RAG and Vector Stores
faiss-gpu>=1.7.2  # Use faiss-cpu if no GPU
pinecone-client>=3.0.0
sentence-transformers>=2.2.0
langchain>=0.1.0
chromadb>=0.4.0

# API and Deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
sse-starlette>=1.8.2

# GGUF Export
gguf>=0.5.0
llama-cpp-python>=0.2.0

# Monitoring and Logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0
loguru>=0.7.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0
omegaconf>=2.3.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0
rich>=13.7.0

# Development Tools
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Optional: Quantization Tools
auto-gptq>=0.5.0
optimum>=1.15.0
